# Remove regex dependency for faster runtime, and compile