--- a/Cargo.toml
+++ b/Cargo.toml
@@ -29,6 +29,7 @@ serde_json = "1.0.48"
 base64 = "0.13.0"
 regex = "1.3.4"
 lazy_static = "1.4.0"
+unicode-id  = "0.3"
 if_chain = "1.0.0"
 scroll = { version = "0.10.1", features = ["derive"], optional = true }