--- a/Cargo.toml
+++ b/Cargo.toml
@@ -27,8 +27,6 @@ url = "2.1.1"
 serde = { version = "1.0.104", features = ["derive"] }
 serde_json = "1.0.48"
 base64 = "0.13.0"
-regex = "1.3.4"
-lazy_static = "1.4.0"
 unicode-id  = "0.3"
 if_chain = "1.0.0"
 scroll = { version = "0.10.1", features = ["derive"], optional = true }