My motivations here are to learn about, and ideally improve the performance of, rust-based frontend tooling.

### Behaviour Changes

The regex crate is used in two separate parts of this lib, **in both cases I have ended up making behavioural changes** - *which I can undo to be more compatible if desired*.

* `filename: &str` -> `index: usize` parsing  in code managing compressed source bundles
  * No longer are any files in the bundle directory ignored, they must all be understood and used by the bundler.
  * I explored a few different options here: https://github.com/willstott101/rust-sourcemap/commit/c6d1debe0df4096c0882859e5b6fa6c9daac693d
* Finding and introspecting JavaScript identifiers
  * `is_valid_javascript_identifier` no longer returns true for invalid identifiers with valid prefixes (see commented test cases)

### unicode-id

I went ahead and used unicode-id as I stole the `is_valid_start` and `is_valid_continue` functions from [swc@5a23949f swc_ecma_ast/src/ident.rs#L180](https://github.com/swc-project/swc/blob/5a23949f1284f71447b9292a374a0cf88759a318/crates/swc_ecma_ast/src/ident.rs#L180) and unicode-id is what swc uses.

### Measurements

|  | Before | After | Notes |
| ------------- | ------------- | - | - |
| Test run (cargo test stdout report) | ~1.0s  | ~0.7s | Bare in mind that I _added_ more tests |
| Compilation from scratch  | ~14.5s  | ~11.5s  | `cargo build --release --examples --offline` |
| `read` example size  | 4770336 | 4761136  | - 9200 bytes   |
| `rewrite` example size  | 4878184  | 4864792  | - 13392 bytes |