@willstott101 2022-11-19
Yeah the use of this identifier stuff is purely within `get_original_function_name`. Trying to use [js-source-scopes](https://github.com/getsentry/js-source-scopes) within here might be a bit confusing though, as that crate appears to depend on this one...