@Swatinem 2022-11-21
yes, IMO this crate should be limited in scope to only sourcemaps. extracting function names (scope information) is a different concern.

I think we can eventually remove that stuff in a breaking release.